+++ StabilityMatrix.Avalonia/Models/ITemplateKey.cs
@@ -1,0 +1,11 @@
+using StabilityMatrix.Avalonia.Controls;
+
+namespace StabilityMatrix.Avalonia.Models;
+
+/// <summary>
+/// Implements a template key for <see cref="DataTemplateSelector"/>
+/// </summary>
+public interface ITemplateKey<out T>
+{
+    T TemplateKey { get; }
+}