+++ StabilityMatrix.Avalonia/ViewModels/Settings/MainSettingsViewModel.cs
@@ -24,7 +24,9 @@ using Avalonia.Threading;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using DynamicData.Binding;
+using ExifLibrary;
 using FluentAvalonia.UI.Controls;
+using MetadataExtractor.Formats.Exif;
 using NLog;
 using SkiaSharp;
 using StabilityMatrix.Avalonia.Animations;
@@ -75,7 +77,8 @@ public partial class MainSettingsViewModel : PageViewModelBase
     public SharedState SharedState { get; }
 
     public override string Title => "Settings";
-    public override IconSource IconSource => new SymbolIconSource { Symbol = Symbol.Settings, IsFilled = true };
+    public override IconSource IconSource =>
+        new SymbolIconSource { Symbol = Symbol.Settings, IsFilled = true };
 
     // ReSharper disable once MemberCanBeMadeStatic.Global
     public string AppVersion =>
@@ -132,7 +135,8 @@ public partial class MainSettingsViewModel : PageViewModelBase
     [ObservableProperty]
     private MemoryInfo memoryInfo;
 
-    private readonly DispatcherTimer hardwareInfoUpdateTimer = new() { Interval = TimeSpan.FromSeconds(2.627) };
+    private readonly DispatcherTimer hardwareInfoUpdateTimer =
+        new() { Interval = TimeSpan.FromSeconds(2.627) };
 
     public Task<CpuInfo> CpuInfoAsync => HardwareHelper.GetCpuInfoAsync();
 
@@ -198,8 +202,16 @@ public partial class MainSettingsViewModel : PageViewModelBase
             true
         );
 
-        settingsManager.RelayPropertyFor(this, vm => vm.SelectedAnimationScale, settings => settings.AnimationScale);
-        settingsManager.RelayPropertyFor(this, vm => vm.HolidayModeSetting, settings => settings.HolidayModeSetting);
+        settingsManager.RelayPropertyFor(
+            this,
+            vm => vm.SelectedAnimationScale,
+            settings => settings.AnimationScale
+        );
+        settingsManager.RelayPropertyFor(
+            this,
+            vm => vm.HolidayModeSetting,
+            settings => settings.HolidayModeSetting
+        );
 
         DebugThrowAsyncExceptionCommand.WithNotificationErrorHandler(notificationService, LogLevel.Warn);
 
@@ -277,16 +289,14 @@ public partial class MainSettingsViewModel : PageViewModelBase
                 CloseButtonText = Resources.Action_RelaunchLater
             };
 
-            Dispatcher
-                .UIThread
-                .InvokeAsync(async () =>
+            Dispatcher.UIThread.InvokeAsync(async () =>
+            {
+                if (await dialog.ShowAsync() == ContentDialogResult.Primary)
                 {
-                    if (await dialog.ShowAsync() == ContentDialogResult.Primary)
-                    {
-                        Process.Start(Compat.AppCurrentPath);
-                        App.Shutdown();
-                    }
-                });
+                    Process.Start(Compat.AppCurrentPath);
+                    App.Shutdown();
+                }
+            });
         }
         else
         {
@@ -313,16 +323,14 @@ public partial class MainSettingsViewModel : PageViewModelBase
     [RelayCommand]
     private void NavigateToSubPage(Type viewModelType)
     {
-        Dispatcher
-            .UIThread
-            .Post(
-                () =>
-                    settingsNavigationService.NavigateTo(
-                        viewModelType,
-                        BetterSlideNavigationTransition.PageSlideFromRight
-                    ),
-                DispatcherPriority.Send
-            );
+        Dispatcher.UIThread.Post(
+            () =>
+                settingsNavigationService.NavigateTo(
+                    viewModelType,
+                    BetterSlideNavigationTransition.PageSlideFromRight
+                ),
+            DispatcherPriority.Send
+        );
     }
 
     #region Package Environment
@@ -350,8 +358,7 @@ public partial class MainSettingsViewModel : PageViewModelBase
         {
             // Save settings
             var newEnvVars = viewModel
-                .EnvVars
-                .Where(kvp => !string.IsNullOrWhiteSpace(kvp.Key))
+                .EnvVars.Where(kvp => !string.IsNullOrWhiteSpace(kvp.Key))
                 .GroupBy(kvp => kvp.Key, StringComparer.Ordinal)
                 .ToDictionary(g => g.Key, g => g.First().Value, StringComparer.Ordinal);
             settingsManager.Transaction(s => s.EnvironmentVariables = newEnvVars);
@@ -405,7 +412,10 @@ public partial class MainSettingsViewModel : PageViewModelBase
 
         await using var _ = new MinimumDelay(200, 300);
 
-        var shortcutDir = new DirectoryPath(Environment.GetFolderPath(Environment.SpecialFolder.StartMenu), "Programs");
+        var shortcutDir = new DirectoryPath(
+            Environment.GetFolderPath(Environment.SpecialFolder.StartMenu),
+            "Programs"
+        );
         var shortcutLink = shortcutDir.JoinFile("Stability Matrix.lnk");
 
         var appPath = Compat.AppCurrentPath;